Move logoutclick to a komponent and add a usersettings page to render on both userscreen and admin page
